refact: split PublicOverview + lazy-loading hooks
